requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
dj-database-url
django-environ
celery
redis
drf-spectacular
cryptography
python-dotenv
django-cors-headers